Correct fishtacos error
Fishtacos.html existed twice, but the wrong one was being updated. The
on that was being updated was in the main folder and not the recipes
folder, so that one was deleted. The one that existed in the recipes
folder was blank, so the code was added to that file properly. I
discovered that my error is if I change file locations but continue to
work on them in code that code will create a new file and ignore the
moved file.